pyproject.toml
[tool.poetry]
name = "doctor-assistant"
version = "0.1.0"
description = ""
authors = ["schonsnatan <natanschons@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.46.1"
langchain = "^0.3.26"
groq = "^0.29.0"
faiss-cpu = "^1.11.0"
python-dotenv = "^1.1.1"
ipykernel = "^6.29.5"
langchain-community = "^0.3.27"
sentence-transformers = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
